+++ planck-c/functions.h
@@ -144,3 +144,6 @@ JSValueRef function_socket_close(JSContextRef ctx, JSObjectRef function, JSObjec
 
 JSValueRef function_sleep(JSContextRef ctx, JSObjectRef function, JSObjectRef thisObject,
                           size_t argc, const JSValueRef args[], JSValueRef *exception);
+
+JSValueRef function_signal_task_complete(JSContextRef ctx, JSObjectRef function, JSObjectRef thisObject,
+                                         size_t argc, const JSValueRef args[], JSValueRef *exception);