+++ planck-c/functions.h
@@ -108,6 +108,9 @@ JSValueRef function_mkdirs(JSContextRef ctx, JSObjectRef function, JSObjectRef t
 JSValueRef function_delete_file(JSContextRef ctx, JSObjectRef function, JSObjectRef thisObject, size_t argc,
                                 const JSValueRef args[], JSValueRef *exception);
 
+JSValueRef function_copy_file(JSContextRef ctx, JSObjectRef function, JSObjectRef thisObject,
+                              size_t argc, const JSValueRef args[], JSValueRef *exception);
+
 JSValueRef function_list_files(JSContextRef ctx, JSObjectRef function, JSObjectRef thisObject, size_t argc,
                                const JSValueRef args[], JSValueRef *exception);