+++ planck-c/io.h
@@ -9,3 +9,5 @@ void write_contents(char *path, char *contents);
 int mkdir_p(char *path);
 
 int mkdir_parents(const char *path);
+
+int copy_file(const char *from, const char *to);