+++ planck-c/globals.h
@@ -1,6 +1,6 @@
 // Global variables used throughout Planck
 
-#define PLANCK_VERSION "2.12.6"
+#define PLANCK_VERSION "2.13.0"
 
 #define EXIT_SUCCESS_INTERNAL -257