+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Fixed
+- `js/clearTimeout` should allow shutdown ([#698](https://github.com/mfikes/planck/issues/698))
 
 ## [2.12.6] - 2018-03-26
 ### Fixed